fix: take add-on attachments into consideration when getting config vars (#133)
* Take add-on attachments into consideration when getting config vars

Now that we support add-ons that are attached from other apps, we need to use the add-on attachment name when determining the AppLink config vars that we need to use.

Previously, we were getting an add-on's info from the Heroku API and using the `config_vars` field. That will work for the app to which the add-on is provisioned to, but not attached apps.

Getting the add-on attachments with `GET /apps/${app}/addon-attachments` allows us to construct the correct config var names, regardless of which app the add-on was provisioned to.

* test: update tests with new addon attachments call

---------

Co-authored-by: Katy Bowman <katy.bowman@salesforce.com>